Refactoring

+++ src/VirtoCommerce.TaskManagement.Web/Controllers/Api/TaskManagementController.cs
@@ -15,6 +15,7 @@ using VirtoCommerce.Platform.Core.Common;
 using VirtoCommerce.Platform.Core.Security;
 using VirtoCommerce.Platform.Core.Settings;
 using VirtoCommerce.TaskManagement.Core;
+using VirtoCommerce.TaskManagement.Core.Extensions;
 using VirtoCommerce.TaskManagement.Core.Models;
 using VirtoCommerce.TaskManagement.Core.Services;
 using VirtoCommerce.TaskManagement.Data.Authorization;
@@ -32,6 +33,7 @@ namespace VirtoCommerce.TaskManagement.Web.Controllers.Api
         private readonly IMemberService _memberService;
         private readonly IMemberSearchService _memberSearchService;
         private readonly ISettingsManager _settingsManager;
+        private readonly IAuthorizationService _authorizationService;
         private readonly MvcNewtonsoftJsonOptions _jsonOptions;
 
         public TaskManagementController(
@@ -40,6 +42,7 @@ namespace VirtoCommerce.TaskManagement.Web.Controllers.Api
             IMemberService memberService,
             IMemberSearchService memberSearchService,
             ISettingsManager settingsManager,
+            IAuthorizationService authorizationService,
             IOptions<MvcNewtonsoftJsonOptions> jsonOptions
             )
         {
@@ -48,22 +51,22 @@ namespace VirtoCommerce.TaskManagement.Web.Controllers.Api
             _memberService = memberService;
             _memberSearchService = memberSearchService;
             _settingsManager = settingsManager;
+            _authorizationService = authorizationService;
             _jsonOptions = jsonOptions.Value;
         }
 
         [HttpGet("{id}")]
         public async Task<ActionResult<WorkTask>> Get(string id, [FromQuery] string respGroup = null)
         {
-            if (!HasPermission(User, TaskPermissions.Read, out var permission))
-            {
-                return Unauthorized();
-            }
-
             var criteria = AbstractTypeFactory<WorkTaskSearchCriteria>.TryCreateInstance();
             criteria.ObjectIds = new[] { id };
             criteria.ResponseGroup = respGroup;
 
-            await FillCriteriaResponsibleInformation(criteria, permission);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, criteria, new TaskAuthorizationRequirement(TaskPermissions.Read));
+            if (!authorizationResult.Succeeded)
+            {
+                return Unauthorized();
+            }
 
             var searchResult = await _workTaskSearchService.SearchAsync(criteria);
             return Ok(searchResult.Results.FirstOrDefault());
@@ -72,12 +75,12 @@ namespace VirtoCommerce.TaskManagement.Web.Controllers.Api
         [HttpPost("search")]
         public async Task<ActionResult<WorkTaskSearchResult>> SearchTasks([FromBody] WorkTaskSearchCriteria criteria)
         {
-            if (!HasPermission(User, TaskPermissions.Read, out var permission))
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, criteria, new TaskAuthorizationRequirement(TaskPermissions.Read));
+            if (!authorizationResult.Succeeded)
             {
                 return Unauthorized();
             }
 
-            await FillCriteriaResponsibleInformation(criteria, permission);
             var result = await _workTaskSearchService.SearchAsync(criteria);
             return Ok(result);
         }
@@ -85,12 +88,12 @@ namespace VirtoCommerce.TaskManagement.Web.Controllers.Api
         [HttpPost("")]
         public async Task<ActionResult<WorkTask>> Create([FromBody] WorkTask workTask)
         {
-            if (!HasPermission(User, TaskPermissions.Create, out _))
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, workTask, new TaskAuthorizationRequirement(TaskPermissions.Create));
+            if (!authorizationResult.Succeeded)
             {
                 return Unauthorized();
             }
 
-            await FillWorkTaskResponsibleInformation(workTask);
             await _workTaskService.SaveChangesAsync(new[] { workTask });
             workTask = await _workTaskService.GetByIdAsync(workTask.Id);
             return workTask;
@@ -99,29 +102,44 @@ namespace VirtoCommerce.TaskManagement.Web.Controllers.Api
         [HttpPut("")]
         public async Task<ActionResult<WorkTask>> Update([FromBody] WorkTask workTask)
         {
-            if (!HasPermission(User, TaskPermissions.Update, out _))
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, workTask, new TaskAuthorizationRequirement(TaskPermissions.Update));
+            if (!authorizationResult.Succeeded)
             {
                 return Unauthorized();
             }
 
-            await FillWorkTaskResponsibleInformation(workTask);
             await _workTaskService.SaveChangesAsync(new[] { workTask });
             return workTask;
         }
 
         [HttpDelete("")]
-        [Authorize(TaskPermissions.Delete)]
         public async Task<ActionResult> Delete([FromBody] string id)
         {
+            var workTask = await _workTaskService.GetByIdAsync(id);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, workTask, new TaskAuthorizationRequirement(TaskPermissions.Delete));
+            if (!authorizationResult.Succeeded)
+            {
+                return Unauthorized();
+            }
+
             await _workTaskService.DeleteAsync(new[] { id });
             return NoContent();
         }
 
         [HttpPost("finish")]
-        [Authorize(TaskPermissions.Finish)]
         public async Task<ActionResult<WorkTask>> Finish([FromQuery] string id, [FromQuery] bool completed, [FromBody] JObject result)
         {
-            var workTask = await _workTaskService.FinishAsync(id, completed, result);
+            var workTask = await _workTaskService.GetByIdAsync(id);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, workTask, new TaskAuthorizationRequirement(TaskPermissions.Finish));
+            if (!authorizationResult.Succeeded)
+            {
+                return Unauthorized();
+            }
+
+            workTask.IsActive = false;
+            workTask.Completed = completed;
+            workTask.Result = result;
+            await _workTaskService.SaveChangesAsync(new[] { workTask });
             return workTask;
         }
 
@@ -169,12 +187,12 @@ namespace VirtoCommerce.TaskManagement.Web.Controllers.Api
             }
             else
             {
-                var assignToMeScope = permission?.AssignedScopes.OfType<TaskAssignToMeScope>().FirstOrDefault();
-                var assignToMyOrganizationScope = permission?.AssignedScopes.OfType<TaskAssignToMyOrganizationScope>().FirstOrDefault();
+                var assignToMeScope = permission?.AssignedScopes.OfType<TaskToMeScope>().FirstOrDefault();
+                var assignToMyOrganizationScope = permission?.AssignedScopes.OfType<TaskToMyOrganizationScope>().FirstOrDefault();
 
                 var memberId = User.FindFirstValue(MemberIdClaimType);
                 var member = await _memberService.GetByIdAsync(memberId);
-                var organizationId = GetMemberOrganizationId(member);
+                var organizationId = member.GetMemberOrganizationId();
 
                 if (permission == null || !permission.AssignedScopes.Any())
                 {
@@ -209,56 +227,6 @@ namespace VirtoCommerce.TaskManagement.Web.Controllers.Api
             var result = setting.AllowedValues.Select(x => new TaskType { Name = x as string }).ToList();
 
             return Ok(result);
-        }
-
-        private async Task FillWorkTaskResponsibleInformation(WorkTask workTask)
-        {
-            if (!string.IsNullOrEmpty(workTask.ResponsibleId))
-            {
-                var member = await _memberService.GetByIdAsync(workTask.ResponsibleId);
-                var organizationId = GetMemberOrganizationId(member);
-
-                workTask.ResponsibleName = member?.Name;
-                workTask.OrganizationId = organizationId;
-            }
-        }
-
-        private async Task FillCriteriaResponsibleInformation(WorkTaskSearchCriteria criteria, Permission permission)
-        {
-            var memberId = User.FindFirstValue(MemberIdClaimType);
-            if (criteria.OnlyAssignedToMe == true)
-            {
-                criteria.ResponsibleIds = new[] { memberId };
-            }
-            else
-            {
-                var assignToMeScope = permission?.AssignedScopes.OfType<TaskAssignToMeScope>().FirstOrDefault();
-                var assignToMyOrganizationScope = permission?.AssignedScopes.OfType<TaskAssignToMyOrganizationScope>()
-                    .FirstOrDefault();
-
-                var member = await _memberService.GetByIdAsync(memberId);
-                var organizationId = GetMemberOrganizationId(member);
-
-                if (!string.IsNullOrEmpty(organizationId) && assignToMyOrganizationScope != null)
-                {
-                    criteria.OrganizationIds = new[] { organizationId };
-                }
-                else if ((assignToMeScope != null || assignToMyOrganizationScope != null) &&
-                         !string.IsNullOrEmpty(memberId))
-                {
-                    criteria.ResponsibleIds = new[] { memberId };
-                }
-            }
-        }
-
-        private static string GetMemberOrganizationId(Member member)
-        {
-            return member?.MemberType switch
-            {
-                nameof(Contact) => (member as Contact)?.Organizations?.FirstOrDefault(),
-                nameof(Employee) => (member as Employee)?.Organizations?.FirstOrDefault(),
-                _ => null
-            };
         }
 
         private bool HasPermission(ClaimsPrincipal user, string permissionName, out Permission permission)
